Fix methods for db

+++ database/model/calc_history.py
@@ -4,10 +4,10 @@ from dataclasses import dataclass
 @dataclass(init=False)
 class History:
     line: str
-    answer: int
+    answer: float
 
 
 @dataclass
 class History_get:
     line: str
-    answer: int
+    answer: float
